Fix names of packages so they're built
This is an example of the big problem with versions:

KERNELRELEASE is read from a variety of sources, which may change at
various stages of the build.  Setting EXTRAVERSION in Makefile can be
used to add '-xenomai-2.6.2.1' to the package name (linux-*-3.5.7 by
default).  Or make-kpkg --append-to-version=-xenomai-2.6.2.1 can be
used, but the resulting /debian/ will not work right.  Setting both
definitely doesn't work.

This hack makes the package names always come out the same, at least.
Before, $(s_package) was linux-source-3.5.7; now it is
linux-source-3.5.7-xenomai-2.6.2.1; and without this, the package is
simply omitted from the build.